+++ docs/changelog.md
@@ -261,6 +261,7 @@ Many feature gates are removed in order to make most features make always availa
 - new macro `slog!`.
 
 ### mem
+- new types: `ArenaBytes`, `ArenaHandle`.
 - move to [base]:
   - macros: `cswap!`, `const_join!`.
   - traits: `MemAligned`.